backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==0.28
python-dotenv==1.0.0
numpy>=1.26.0  # Phiên bản hỗ trợ Python 3.13
pydantic>=2.9.2,<3.0
aiofiles==23.2.1
python-multipart>=0.0.20
pdfplumber==0.9.0